refac: Move tool execution functions into tools module
