+++ src/tools/mod.rs
@@ -2,7 +2,10 @@ pub mod install;
 pub mod list;
 pub mod pip;
 mod pixi_config;
+mod run;
 pub mod tools;
 pub mod uninstall;
 pub mod utils;
 pub mod uv;
+
+pub use run::run_tool_binary;
